FPAD-8430: Separate interventions processor handler lambda and logic into dedicated files
